+++ app/src/main/java/com/tomaszrykala/wordduel/main/ui/GameBoardScreen.kt
@@ -28,7 +28,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.tomaszrykala.wordduel.game.board.BoardTile
 import com.tomaszrykala.wordduel.R
 import com.tomaszrykala.wordduel.game.keyboard.KeyTile
 import com.tomaszrykala.wordduel.game.state.GameState
